+++ build_doc.py
@@ -4,6 +4,8 @@ Apply an arbitrary filter to each docstring
 '''
 
 import pdoc
+from io import StringIO
+from contextlib import redirect_stdout
 
 substitutions = [
 	('δ13C_VPDB', 'δ<sup>13</sup>C<sub>VPDB</sub>'),
@@ -49,18 +51,28 @@ def myfilter(docstr):
 			work[k] = '`'.join(work[k])
 	return ('```'.join(work))
 
+for code_input, code_output in [
+	('code_examples/virtual_data/example.py', 'code_examples/virtual_data/output.txt'),
+	('code_examples/data_quality/example.py', 'code_examples/data_quality/output.txt'),
+	]:
+
+	with open(code_input) as fid:
+		code = fid.read()
+
+	f = StringIO()
+	with redirect_stdout(f):
+		exec(code)
+
+	with open(code_output, 'w') as fid:
+		fid.write(f.getvalue())
+
+	if 'data_quality' in code_input:
+		data47.plot_distribution_of_analyses(dir = 'docs', filename = 'time_distribution.png', dpi = 120)
+		data47.plot_sessions(dir = 'docs', filetype = 'png')
+		data47.plot_residuals(dir = 'docs', filename = 'residuals.png', kde = True)
+
 pdoc.render.env.filters['myfilter'] = myfilter
 pdoc.render.configure(template_directory = 'pdoc_templates')
 
 with open('docs/index.html', 'w') as fid:
 	fid.write(pdoc.pdoc('D47crunch'))
-
-# foo = '''
-# Create foo δ13C bar Δ47 δ18O_VSMOW `δ13C_VPDB`, `Δ48`.
-# 
-# ````py
-# Create foo δ13C bar `Δ47` δ18O_VSMOW.
-# ```
-# '''
-# 
-# print(myfilter(foo))