+++ changelog.md
@@ -1,5 +1,38 @@
 # Changelog
 
+## v2.2.0
+*Released on 2023-07-20*
+
+### Command-line interface (CLI)
+* Rejoice, you no longer need to know Python: it is now possible to process a multi-session Δ47 dataset with custom anchors and custom UID/sample exclusion list by simply calling `D47crunch -a anchors.csv -e exclude.csv -o outdir rawdata.csv`.
+
+### New features
+* Add `yspan` option to `D4xdata.plot_residuals()`
+* Added `shuffle` option to `virtual_data()`.
+* Added `filetype` option to `D4xdata.plot_sessions()`.
+* Added `dpi` option to `D4xdata.plot_sessions()`, `D4xdata.plot_residuals()`.
+
+### Bugfix
+* Fix error in `D4xdata.plot_residuals()` when `hist = False` and `kde = False`.
+
+## v2.1.1
+*Released on 2023-05-16*
+
+### Bugfix
+* `D4xdata.compute_r()` uses an improved computation for degrees of freedom for arbitrary subsets of sessions and/or samples, yielding more estimates of analytical repeatabilities for Δ47 and Δ48.
+
+### New feature
+* Added `kde` option to `D4xdata.plot_residuals()`
+
+### Other changes
+* Minor improvement to y axis tick labels) in `D4xdata.plot_residuals()`.
+
+## v2.1.0
+*Released on 2023-05-14*
+
+### New feature
+* `D4xdata.plot_bulk_compositions()` plots the dispersion of δ13C and δ18O values for each sample.
+
 ## v2.0.6
 *Released on 2023-05-13*